ecocloud_env/server/requirements.txt
openenv-core
fastapi
uvicorn
pydantic>=2.0
